refactors add new reminder to use single-reminder component
