+++ app/components/single-reminder.js
@@ -16,5 +16,15 @@ export default Ember.Component.extend({
         reminder => reminder.set('isEditing', false)
       );
     },
+
+    createReminder() {
+      const reminder = this.getProperties('title', 'date', 'notes');
+      console.log(this);
+      console.log(reminder);
+      reminder.date = new Date(reminder.date);
+      this.get('store').createRecord('reminder', reminder).save().then(() => {
+        this.setProperties({ title: '', date: '', notes: '' });
+      });
+    },
   }
 });
