+++ app/routes/reminders/new.js
@@ -1,0 +1,10 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  model() {
+    return {
+      isEditing: true,
+      isNew: true
+    };
+  }
+});
