--- app/templates/components/add-reminder.hbs
@@ -1,17 +1,0 @@
-<h2> New Reminder </h2>
-
-<form class="add-reminder--form" {{action 'createReminder' on="submit"}}>
-  <label>
-    Title
-    {{input class="spec-input-title" value=title}}
-  </label>
-  <label>
-    Date
-    {{input type="date" class="spec-input-date" value=date}}
-  </label>
-  <label>
-    Notes
-    {{textarea class="spec-textarea-notes" value=notes}}
-  </label>
-  <button class="add-reminder--submit">Submit</button>
-</form>
