+++ app/templates/reminders/new.hbs
@@ -1,1 +1,5 @@
-{{add-reminder}}
+<h1>New Reminder</h1>
+{{single-reminder
+  reminder=model
+  model=model
+}}
