+++ app/templates/reminders/reminder.hbs
@@ -1,3 +1,6 @@
-{{single-reminder reminder=reminder model=model}}
+{{single-reminder
+  reminder=reminder
+  model=model
+}}
 
 {{outlet}}
